+++ derive/CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Made `Event = ...` attribute support actual event type and not just
+  string representation of it
 - Bumped minimum required Rust version to `1.34.0`